backend/requirements.txt
fastapi
uvicorn
pillow
torch
torchvision
requests
python-multipart